requirements.txt
openai
httpx
pydantic
fastapi
python-dotenv
langchain_community
langchain_openai
uvicorn
langchain
tiktoken>=0.6.0
firebase_admin
numpy
selenium
beautifulsoup4
